protocol_analyzer/requirements.txt
# Core (always required)
python-dotenv>=1.0.0

# Install only the SDKs for the providers you are using:
openai>=1.30.0              # Required for OpenAI AND DeepSeek (uses same SDK)
anthropic>=0.25.0           # Required for Claude
google-generativeai>=0.5.0  # Required for Gemini
